refactor: Create controller
